# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "This is the API documentation for my Invoices application.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/invoice/list": {
            "post": {
                "tags": [
                    "Invoice"
                ],
                "summary": "List Invoices",
                "description": "Returns a list of invoices based on filtering, sorting, and pagination options. Requires authentication.",
                "operationId": "1589492b3a0664f87658f0fde201b580",
                "requestBody": {
                    "description": "Filtering, sorting, and pagination parameters",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaginatedListFilter"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of invoices",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Invoice"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/invoice/{id}": {
            "get": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Get Invoice",
                "description": "Retrieve an invoice by its ID. Requires authentication.",
                "operationId": "ea794c8902eebd45c41a2be4e2a9daad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoice found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error details..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Update Invoice",
                "description": "Updates an existing invoice. Requires authentication.",
                "operationId": "71da025873fb8b7132e9d80079585b91",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated invoice data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Invoice"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invoice updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": "{ 'field': ['Error message'] }"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Delete Invoice",
                "description": "Deletes an invoice by its ID. Requires authentication.",
                "operationId": "7a5671d649c49132a2a7fd413c261ed2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoice deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invoice deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invoice not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invoice not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/invoice": {
            "post": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Create Invoice",
                "description": "Creates a new invoice. Requires authentication.",
                "operationId": "ce42c64432fe3c0715e3b4572154dd67",
                "requestBody": {
                    "description": "Invoice data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Invoice"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Invoice created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": "{ 'field': ['Error message'] }"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Filter": {
                "title": "Filter",
                "description": "A single filter condition.",
                "required": [
                    "field",
                    "operator"
                ],
                "properties": {
                    "field": {
                        "description": "Field name to filter on.",
                        "type": "string",
                        "example": "user_id"
                    },
                    "value": {
                        "description": "Value to filter by.",
                        "type": "string",
                        "example": "36"
                    },
                    "fieldType": {
                        "description": "Type of the field.",
                        "type": "string",
                        "enum": [
                            "text",
                            "date",
                            "number",
                            "boolean",
                            "set",
                            "array"
                        ],
                        "example": "number"
                    },
                    "operator": {
                        "description": "Operator for the filter condition.",
                        "type": "string",
                        "enum": [
                            "contains",
                            "notContains",
                            "equals",
                            "notEqual",
                            "startsWith",
                            "endsWith",
                            "blank",
                            "notBlank",
                            "greaterThan",
                            "greaterThanOrEqual",
                            "lessThan",
                            "lessThanOrEqual",
                            "inRange"
                        ],
                        "example": "equals"
                    },
                    "rangeValue": {
                        "description": "Secondary value for range filters (if applicable).",
                        "type": "string",
                        "example": "50",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginatedListFilter": {
                "title": "Paginated List Filter",
                "description": "Parameters for filtering, sorting, and pagination.",
                "required": [
                    "filters",
                    "sort",
                    "limit"
                ],
                "properties": {
                    "filters": {
                        "description": "Array of filter objects",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Filter"
                        }
                    },
                    "sort": {
                        "$ref": "#/components/schemas/Sort"
                    },
                    "limit": {
                        "description": "Maximum number of records to return",
                        "type": "integer",
                        "example": 15
                    }
                },
                "type": "object"
            },
            "Sort": {
                "title": "Sort",
                "description": "Sorting options for a query.",
                "required": [
                    "field",
                    "direction"
                ],
                "properties": {
                    "field": {
                        "description": "Field to sort by.",
                        "type": "string",
                        "example": "invoice_number"
                    },
                    "direction": {
                        "description": "Sort direction.",
                        "type": "string",
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "example": "asc"
                    }
                },
                "type": "object"
            },
            "Invoice": {
                "title": "Invoice",
                "description": "Invoice model representing an invoice issued to a user",
                "required": [
                    "user_id",
                    "tax_profile_id",
                    "invoice_number",
                    "invoice_date",
                    "total_amount",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the invoice",
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "user_id": {
                        "description": "ID of the user who owns the invoice",
                        "type": "integer",
                        "example": 1
                    },
                    "tax_profile_id": {
                        "description": "ID of the tax profile associated with the invoice",
                        "type": "integer",
                        "example": 2
                    },
                    "invoice_number": {
                        "description": "Invoice number",
                        "type": "string",
                        "example": "INV-1001"
                    },
                    "description": {
                        "description": "Description of the invoice",
                        "type": "string",
                        "example": "Invoice for consulting services rendered"
                    },
                    "invoice_date": {
                        "description": "Date of the invoice",
                        "type": "string",
                        "format": "date",
                        "example": "2021-12-31"
                    },
                    "total_amount": {
                        "description": "Total amount for the invoice",
                        "type": "number",
                        "format": "float",
                        "example": 1234.56
                    },
                    "status": {
                        "description": "Current status of the invoice",
                        "type": "string",
                        "example": "pending"
                    },
                    "created_at": {
                        "description": "Timestamp when the invoice was created",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2022-01-01T12:00:00Z",
                        "nullable": true
                    },
                    "updated_at": {
                        "description": "Timestamp when the invoice was last updated",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2022-01-02T12:00:00Z",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}